// utils/types.ts
export interface TreeNode {
  name: string;
  value?: number;
  children: TreeNode[];
}
 export interface ExcelRow {
  Source: string;
  Account: string;
  Year: number;
  TxnType: string;
  Tag: string;
  Amount: number;
}

export interface TreeNode {
  name: string;
  value?: number;
  children: TreeNode[];
}